Functions used with local storage added to helpers
