+++ helpers.js
@@ -1,3 +1,8 @@
+const Keys = {
+    savedEmail: 'savedEmail',
+    savedPassword: 'savedPassword',
+}
+
 export const createNewElement = (tag) => {
     return document.createElement(tag);
 };
@@ -23,4 +28,20 @@ export const validateForm = (email, password) => {
         return false;
     }
     return true;
+}
+
+export const saveInputValue = (email, password) => {
+    localStorage.setItem(Keys.savedEmail, email);
+    localStorage.setItem(Keys.savedPassword, password);
+}
+
+export const loadInputValues = () => {
+    const savedEmail = localStorage.getItem(Keys.savedEmail);
+    const savedPassword = localStorage.getItem(Keys.savedPassword);
+    return { savedEmail, savedPassword };
+};
+
+export const deleteInputValue = () => {
+    localStorage.removeItem(Keys.savedEmail);
+    localStorage.removeItem(Keys.savedPassword);
 }
