+++ index.html
@@ -9,7 +9,7 @@
 
 <body>
     <div id="root"></div>
-    <script src="form.js" type="module"></script>
+    <script src="./dist/bundle.js"></script>
 </body>
 
 </html>
